requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0

# Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0

# Configuration
pyyaml>=6.0

# Utilities
psutil>=5.9.0

# Optional: Fisher visualization
# scipy>=1.11.0
# scikit-learn>=1.3.0
